+++ src/components/autocomplete/script.ts
@@ -11,6 +11,7 @@ import StDropdownOption from '../dropdown-option/index.vue';
 import StDropdown from '../dropdown/index.vue';
 import StDropdownScript from '../dropdown/script';
 import { DropdownBindProperties } from '../dropdown/types';
+import StInput from '../input/script';
 import {
   PopperPlacement,
   TriggerType,
@@ -231,5 +232,13 @@ export default class StAutocomplete extends Vue {
 
   onDocClick() {
     this.togglePopperVisibility(false);
+  }
+
+  focus(): void {
+    (this.$refs.input as StInput).focusInput();
+  }
+
+  blur(): void {
+    (this.$refs.input as StInput).blurInput();
   }
 }